Fixes caching and routing issues with dots in wiki ids
Monkey patched ActionController::Caching::Actions:ActionCachePath#extract_session to never include an extension/format when caching.
I think this should be OK since the only format for wiki pages is HTML -no xls, js, or xml output-

Fixed the routes for any wiki action with an id to match ids on anything but a '/'. This works because '/' in a URL is escaped as %2F.
There is 1 test that fails. WikiControllerTest#test_edit_page_with_special_symbols writes a wiki page with a '/' in its name -this is fine since it will get escaped-,
however TestProcess#process does not escape the id correctly. Wrapping the id with URI.escape will not work either since it never gets unescaped.